requirements.txt
line-bot-sdk==2.3.0
flask==2.1.2
requests==2.28.0
pandas==2.0.3
lxml==4.9.3
bs4==0.0.1